Module relay_common_actors::controller

Defines an actor to control system run and shutdown.

See the Controller struct for more information.

Structs

Configure

Configures the Controller with new parameters.

Controller

Actor to start and gracefully stop an actix system.

Shutdown

Shutdown request message sent by the Controller to subscribed actors.

Subscribe

Subscribtion message for Shutdown events.